WIP - proper quote "Executing task" message
The "Executing task" messag is followed by what looks
like a command line. However, it is not safe to cut-and-paste
this string into a shell because it is not quoted properly.

Eg. If the 1 argument task below is reported, it would become
a 2 argument task when cut-and-pasted because the argument is
not quoted:

echo "two words"

is reported as:

Executing task: echo two words

The correct message should be:

Executing task: echo "two words"

WIP because the quoting is done with @mergesium/shell-quote
package, but dependencies are not allowed. I will figure out how
to bring the function to vscode later.

see https://github.com/microsoft/vscode/wiki/Source-Code-Organization